+++ src/thrust/model.cmake
@@ -18,6 +18,9 @@ register_flag_optional(BACKEND
         "
         "CUDA")
 
+      register_flag_optional(MANAGED "Enabled managed memory mode."
+        "OFF")
+
 register_flag_optional(CMAKE_CUDA_COMPILER
         "[THRUST_IMPL==CUDA] Path to the CUDA nvcc compiler"
         "")
@@ -34,6 +37,9 @@ register_flag_optional(CUDA_EXTRA_FLAGS
 
 macro(setup)
     set(CMAKE_CXX_STANDARD 14)
+    if (MANAGED)
+      register_definitions(MANAGED)
+    endif ()
 
     if (${THRUST_IMPL} STREQUAL "CUDA")